版本迭代完成

+++ lib/utils/time_util.dart
@@ -254,7 +254,7 @@ class TimeUtil {
   }
 
   /// 金额 分转元 11000 = 110.00
-  static String amountConversion(int amount) {
+  static String amountConversion(double amount) {
     var minutes = (amount / 100).toStringAsFixed(2);
     return minutes;
   }
